requirements.txt

ipython
attrs
